Initial Commit
Includes the maze code for the first two levels, Conway’s Game of Life
in the end, and the beginning of the Hangman code which I have just
started coding.

After talking with Kiang and discussing what my project will encompass,
I have decided to just start working on the Hangman code, and I have
realized that it is harder than I thought it would be.