Work Log #9
CJ helped me get the Wordplay method to correctly add words from
words.txt to the words arraylist. I just had to put words.txt in the
same folder as my other code.

I fixed a problem with the letterToIndex method that was getting an out
of bounds error.

The biggest change I made was refreshing the words arraylist. Now, the
program will check and remove any words that do not match the
information known. For example, if the letter a was guessed and is not
in the word, all the words in the array list that have the letter a
will be removed.