Small Changes
I realized that I forgot to clear the arraylists for words and wrong
letters, which was a problem when Aliya tried to play Hangman for the
second tie. I fixed that by clearing the array lists in the start()
method.

I spent class beta-testing Nate’s project, having my Hangman maze beta
tested by Nate and others, and helping people with their code.

I’m satisfied with my project as it is, but I have a few ideas on how
to add to it:

1. Add more mazes - There are only two mazes right now.

2. Have the Hangman program change in difficulty - I was originally
going to have the number of guesses start at 4 or 5, and increase by 1
for each level to make the game more difficult.